Extract VLMEvalKit adapter from run_easi_eval.py
Move VLMEvalKit-specific constants, dataset preparation, verification,
and progress monitoring functions into backends/vlmevalkit.py. The
VLMEvalKitAdapter class implements the BackendAdapter interface with
build_cmd, prepare_datasets, poll_progress, detect_completion,
get_result_files, extract_scores, and get_env_overrides methods.

run_easi_eval.py now imports these from the adapter module while
keeping EASI_8, EXTRA, display classes, and main() unchanged.
